requirements.txt
openai>=1.0.0
datasets>=2.0.0
python-dotenv>=1.0.0
sacrebleu>=2.0.0
pandas
numpy
matplotlib
networkx
openpyxl
reportlab
